+++ lib/core/constants/color.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 
-const kAccentColor = Color(0xFFFF385C);
 
 //아래의 주 색상들은 어떤 색 사용할지 정해서 주로 쓰는 색으로 사용 고고곡
 
@@ -15,3 +14,6 @@ const INPUT_BG_COLOR = Color(0xFFFBFBFB);
 
 //텍스트 필드 테두리 색상
 const INPUT_BORDER_COLOR = Color(0xFFF3F2F2);
+
+//에러 메시지 색상
+const ERROR_COLOR = Color(0xFFFF385C);
