+++ lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:bookbox/core/constants/styles.dart';
 import 'package:bookbox/ui/main/main_page.dart';
-import 'package:bookbox/ui/user/login_page.dart';
+import 'package:bookbox/ui/user/login/login_page.dart';
 import 'package:flutter/material.dart';
 
 void main() {
